requirements.txt
asgiref==3.8.1
Django==5.1.4
django-admin-tools==0.9.3
django-excel-response2==3.0.6
django-grappelli==4.0.1
django-six==1.0.5
et_xmlfile==2.0.0
excel-base==1.0.4
isoweek==1.3.3
openpyxl==3.1.5
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.17.0
sqlparse==0.5.2
TimeConvert==3.0.13
typing_extensions==4.12.2
tzlocal==5.2
xlwt==1.3.0
